Initial implementation of /dev/input/eventX monitoring
This implements a multithreaded (parallel subshell) script event_detect.sh
which monitors the event device ids corresponding to the names
selected in the dev_input_devices variable of the idle_detect.conf file.

This script must be run as root to have proper access to the event devices.
The script creates (if necessary) and writes the dat files for each selected
device to the /run/event_detect directory. Each file contains the line count
of the streaming output from the evemu-record for that id.

Note that this script does NOT re-source the config file currently if the
list of devices have changed, nor does it detect if a device was inserted, which
could change the device ids.