Initial skeleton for idle_detect C++ version
Note this skeleton was the result of a conversation with
Gemini Advanced 2.0 Flash Thinking to write a C++ program
following the style of event_detect, including access to XSS module
and Wayland APIs for KDE and Gnome. It will have to be
seriously modified to actually work.

The cmake build files have been updated for two targets now.