Refine the logic for sending the last_active_time to event_detect
While the user is active, the last_active_time only needs to be
sent to event_detect if the last_active_time has actually changed.

This reduces the number of messages. Consider a person who has some
activity and then stops. The prior logic would continue to send
messages to event_detect until the trigger threshold for idle is
reached, finally ceasing when idle logic is triggered. Here as soon
as the user stops activity, the idle_time will increase, so the
effective last_active_time will then become constant at the last
time the user actually generated activity. So the last message
that will be sent to event_detect will coincide with this time,
until new activity occurs.