Change shmem output to contain int64_t[2] with current timestamp and last active time
This is to conform with BOINC side spec for reading last active.

The int64_t[2] has two elements:

0: The current timestamp (i.e. at the time of posting)
1: The last active time

This allows BOINC (or other DC systems) to monitor the shmem segment
for a heartbeat.

Note the name of the shared memory segment has been changed to
/idle_detect_shmem

The small utility for the command line's name has been changed to
read_shmem_timestmps and outputs the two times on one line separated
by a space, which is ideal for scripting use or testing.