Remove atomic from shmem. Implement mutex for local writer locking
This controls concurrency on the event_detect writer side. The reader
takes the risk of a torn read. Note that an architecural review of
this, there is concurrence due to the low update frequency that
the chances of a torn read are low, and atomics can't be implemented
on C++03 code on some DC systems, and the overhead of using
pthread_mutex_t is not worth it.